[ecore][go] remove abstractnotification pointer from all local notifications ( basicnotifyinglist / basicestorelist / eresourceimpl )
